+++ src/pgcopylib/__init__.py
@@ -1,20 +1,28 @@
 """PGCopy bynary dump parser."""
 
 from .common import (
+    PGCopyError,
     PGCopyRecordError,
     PGCopySignatureError,
+    PGCopyTypeError,
+    PGCopyValueError,
     PGOid,
 )
+from .core.dtype import PostgreSQLDtype
 from .reader import PGCopyReader
 from .writer import PGCopyWriter
 
 
 __all__ = (
+    "PGCopyError",
     "PGCopyReader",
     "PGCopyRecordError",
     "PGCopySignatureError",
+    "PGCopyTypeError",
+    "PGCopyValueError",
     "PGCopyWriter",
     "PGOid",
+    "PostgreSQLDtype",
 )
 __author__ = "0xMihalich"
-__version__ = "0.2.4.dev0"
+__version__ = "0.2.4.dev1"