+++ src/pgcopylib/common/__init__.py
@@ -1,18 +1,34 @@
+"""Common PGCopy classes, modules, functions and associate dictionaries."""
+
+from . import signatures as Signature
 from .enums import (
     ArrayOidToOid,
     PGOid,
     PGOidToDType,
 )
-from .errors import (
+from ..core.errors import (
+    PGCopyError,
     PGCopyRecordError,
     PGCopySignatureError,
+    PGCopyTypeError,
+    PGCopyValueError,
+)
+from .repr import (
+    pgcopylib_repr,
+    table_repr,
 )
 
 
 __all__ = (
     "ArrayOidToOid",
+    "PGCopyError",
     "PGCopyRecordError",
     "PGCopySignatureError",
+    "PGCopyTypeError",
+    "PGCopyValueError",
     "PGOid",
     "PGOidToDType",
+    "Signature",
+    "pgcopylib_repr",
+    "table_repr",
 )