+++ src/pgcopylib/common/errors.py
@@ -1,6 +1,0 @@
-class PGCopySignatureError(ValueError):
-    """Signature not match."""
-
-
-class PGCopyRecordError(ValueError):
-    """Record length error."""