+++ src/pgcopylib/reader.py
@@ -1,81 +1,107 @@
+from collections.abc import Generator
 from io import (
     BufferedReader,
     BytesIO,
 )
-from typing import (
-    Any,
-    Generator,
-    TYPE_CHECKING,
+from types import FunctionType
+from typing import Any
+from struct import (
+    error as UnpackError,
+    unpack,
 )
-from struct import unpack
 
 from .common import (
     ArrayOidToOid,
     PGOid,
     PGOidToDType,
     PGCopySignatureError,
+    Signature,
+    pgcopylib_repr,
 )
-from .common.base import (
+from .core.functions import (
     read_num_columns,
     read_record,
     skip_all,
 )
-from .common.dtypes import PostgreSQLDtype
-
-if TYPE_CHECKING:
-    from types import FunctionType
+from .core.dtype import PostgreSQLDtype
 
 
 class PGCopyReader:
     """PGCopy dump reader."""
 
+    fileobj: BufferedReader
+    pgtypes: list[PGOid]
+    pgoid: list[int]
+    pgoid_functions: list[FunctionType]
+    postgres_dtype: list[PostgreSQLDtype]
+    header: bytes
+    flags_area: list[int]
+    is_oid_enable: bool
+    is_empty: bool
+    column_length: int
+    header_ext_length: int
+    num_columns: int
+    num_rows: int
+    buffer_object: BytesIO
+
     def __init__(
         self,
-        file: BufferedReader,
-        pgtypes: list[PGOid] = [],
+        fileobj: BufferedReader,
+        pgtypes: list[PGOid] | None = None,
     ) -> None:
         """Class initialization."""
 
-        self.file = file
+        if not pgtypes:
+            pgtypes = []
+
+        self.fileobj = fileobj
         self.pgtypes = pgtypes
+        self.header = self.fileobj.read(11)
 
-        self.header: bytes = self.file.read(11)
+        if self.header != Signature.HEADER:
+            raise PGCopySignatureError("PGCopy signature not match!")
 
-        if self.header != b"PGCOPY\n\xff\r\n\x00":
-            msg = "PGCopy signature not match!"
-            raise PGCopySignatureError(msg)
-
-        self.flags_area: list[int] = [
+        self.flags_area = [
             (byte >> i) & 1
-            for byte in self.file.read(4)
+            for byte in self.fileobj.read(4)
             for i in range(7, -1, -1)
         ]
-        self.is_oid_enable: bool = bool(self.flags_area[16])
-        self.header_ext_length: int = unpack("!i", self.file.read(4))[0]
+        self.is_oid_enable = bool(self.flags_area[16])
 
         if self.is_oid_enable:
-            self.column_length: int = 6
+            self.column_length = 6
         else:
-            self.column_length: int = 2
+            self.column_length = 2
 
-        self.num_columns: int = read_num_columns(
-            self.file,
-            self.column_length,
-        )
-        self.num_rows: int = 0
-        self.postgres_dtype: list[PostgreSQLDtype] = [
+        try:
+            self.header_ext_length = unpack(
+                "!i",
+                self.fileobj.read(4),
+            )[0]
+            self.num_columns = read_num_columns(
+                self.fileobj,
+                self.column_length,
+            )
+            self.is_empty = False
+        except UnpackError:
+            self.header_ext_length = 0
+            self.num_columns = len(self.pgtypes)
+            self.is_empty = True
+
+        self.num_rows = 0
+        self.postgres_dtype = [
             PGOidToDType[self.pgtypes[column]]
             if self.pgtypes else PostgreSQLDtype.Bytes
             for column in range(self.num_columns)
         ]
-        self.pgoid_functions: list[FunctionType] = [
+        self.pgoid_functions = [
             PGOidToDType[ArrayOidToOid[self.pgtypes[column]]].read
             if self.pgtypes and ArrayOidToOid.get(
                 self.pgtypes[column]
             ) else None
             for column in range(self.num_columns)
         ]
-        self.pgoid: list[int] = [
+        self.pgoid = [
             ArrayOidToOid[self.pgtypes[column]].value
             if self.pgtypes and ArrayOidToOid.get(
                 self.pgtypes[column]
@@ -84,12 +110,12 @@ class PGCopyReader:
         ]
         self.buffer_object = BytesIO()
 
-    def __count_rows(self) -> None:
-        """Count rows."""
+    def read_info(self) -> None:
+        """Read info without reading data."""
 
         if not self.num_rows:
             self.num_rows = skip_all(
-                self.file,
+                self.fileobj,
                 self.column_length,
                 self.num_columns,
                 self.num_rows,
@@ -98,13 +124,16 @@ class PGCopyReader:
     def read_row(self) -> Generator[Any, None, None]:
         """Read single row."""
 
+        if self.is_empty:
+            return
+
         for postgres_dtype, pgoid_function, pgoid in zip(
             self.postgres_dtype,
             self.pgoid_functions,
             self.pgoid,
         ):
             yield read_record(
-                self.file,
+                self.fileobj,
                 postgres_dtype.read,
                 pgoid_function,
                 self.buffer_object,
@@ -116,67 +145,31 @@ class PGCopyReader:
 
         columns = self.num_columns
 
-        while columns != 0xffff:
+        while columns != 0xffff and not self.is_empty:
             yield [*self.read_row()]
             self.num_rows += 1
             columns = read_num_columns(
-                self.file,
+                self.fileobj,
                 self.column_length,
             )
 
     def tell(self) -> int:
         """Return current position."""
 
-        return self.file.tell()
+        return self.fileobj.tell()
 
     def close(self) -> None:
         """Close file object."""
 
-        self.file.close()
+        if hasattr(self.fileobj, "close"):
+            self.fileobj.close()
 
     def __repr__(self) -> str:
-        """PGCopy info in interpreter."""
+        """String representation of PGCopyReader."""
 
-        return self.__str__()
-
-    def __str__(self) -> str:
-        """PGCopy info."""
-
-        self.__count_rows()
-
-        if not self.pgtypes:
-            self.pgtypes = [
-                PGOid.bytea
-                for _ in range(self.num_columns)
-            ]
-
-        def to_col(text: str) -> str:
-            """Format string element."""
-
-            text = text[:14] + "…" if len(text) > 15 else text
-            return f" {text: <15} "
-
-        empty_line = (
-            "├─────────────────┼─────────────────┤"
+        return pgcopylib_repr(
+            self.pgtypes or [PGOid.bytea for _ in range(self.num_columns)],
+            self.num_columns,
+            self.num_rows,
+            "reader",
         )
-        end_line = (
-            "└─────────────────┴─────────────────┘"
-        )
-        _str = [
-            "<PGCopy dump reader>",
-            "┌─────────────────┬─────────────────┐",
-            "│ Column Number   │ PostgreSQL Type │",
-            "╞═════════════════╪═════════════════╡",
-        ]
-
-        for column, pgtype in zip(range(self.num_columns), self.pgtypes):
-            _str.append(
-                f"│{to_col(f'Column_{column}')}│{to_col(pgtype.name)}│",
-            )
-            _str.append(empty_line)
-
-        _str[-1] = end_line
-        return "\n".join(_str) + f"""
-Total columns: {self.num_columns}
-Total rows: {self.num_rows}
-"""